+++ module/actes-generique/action/IParapheurRecup.php
@@ -58,7 +58,7 @@ class IParapheurRecup extends ActionExecutor
 
         $actes = $this->getDonneesFormulaire();
         $dossierID = $actes->get('iparapheur_dossier_id');
-        $info = $signature->getSignature($dossierID, false);
+        $info = $signature->getSignature($dossierID);
         if (! $info) {
             $this->setLastMessage("La signature n'a pas pu être récupérée : " . $signature->getLastError());
             return false;