+++ module/commande-generique/action/FournisseurCommandeReceptionParapheur.php
@@ -2,6 +2,8 @@
 
 class FournisseurCommandeReceptionParapheur extends ActionExecutor
 {
+    private const string ACTION_NAME_REJECT = 'rejet-iparapheur';
+
     /**
      * @return bool
      * @throws Exception
@@ -16,42 +18,57 @@ class FournisseurCommandeReceptionParapheur extends ActionExecutor
 
         $dossierID = $donneesFormulaire->get('iparapheur_dossier_id');
 
+        $error = false;
         $all_historique = false;
         try {
             $all_historique = $signature->getAllHistoriqueInfo($dossierID);
+            if (! $all_historique) {
+                $error = 'La connexion avec le iParapheur a échoué : ' . $signature->getLastError();
+            }
         } catch (Exception $e) {
-            $this->traitementErreur($signature, $e->getMessage());
+            $error = $e->getMessage();
         }
 
-        if (! $all_historique) {
-            $message = 'La connexion avec le iParapheur a échoué : ' . $signature->getLastError();
-            $this->traitementErreur($signature, $message);
-            return false;
+        $lastHistorique = false;
+        $lastCompletedHistorique = false;
+        if (!$error) {
+            $array2XML = new Array2XML();
+            $historique_xml = $array2XML->getXML(
+                'iparapheur_historique',
+                json_decode(
+                    json_encode($all_historique, JSON_THROW_ON_ERROR),
+                    true,
+                    512,
+                    JSON_THROW_ON_ERROR
+                )
+            );
+            $donneesFormulaire->setData('has_historique', true);
+            $donneesFormulaire->addFileFromData(
+                'iparapheur_historique',
+                'iparapheur_historique.xml',
+                $historique_xml
+            );
+            $lastHistorique = $signature->getLastHistorique($all_historique);
+            $lastCompletedHistorique = $signature->getLastCompletedHistorique($all_historique);
+            $donneesFormulaire->setData('parapheur_last_message', $lastCompletedHistorique);
         }
-
-        $array2XML = new Array2XML();
-        $historique_xml = $array2XML->getXML(
-            'iparapheur_historique',
-            json_decode(json_encode($all_historique), true)
-        );
-
-        $donneesFormulaire->setData('has_historique', true);
-        $donneesFormulaire->addFileFromData('iparapheur_historique', 'iparapheur_historique.xml', $historique_xml);
-
-        $lastHistorique = $signature->getLastHistorique($all_historique);
-        $lastCompletedHistorique = $signature->getLastCompletedHistorique($all_historique);
-        $donneesFormulaire->setData('parapheur_last_message', $lastCompletedHistorique);
 
         if ($signature->isFinalState($lastHistorique)) {
             return $this->retrieveDossier($dossierID);
         }
         if ($signature->isRejected($lastHistorique)) {
-            $this->rejeteDossier($dossierID, $lastCompletedHistorique);
-        } else {
-            $this->traitementErreur($signature, $lastHistorique);
+            return $this->rejeteDossier($dossierID, $lastCompletedHistorique);
         }
-        $this->setLastMessage($lastCompletedHistorique);
-        return true;
+
+        $this->traitementErreur($signature, $lastHistorique);
+
+
+        if (!$error) {
+            $this->setLastMessage($lastCompletedHistorique);
+            return true;
+        }
+        $this->setLastMessage($error);
+        return false;
     }
 
     /**
@@ -75,7 +92,7 @@ class FournisseurCommandeReceptionParapheur extends ActionExecutor
     /**
      * @throws Exception
      */
-    public function rejeteDossier($dossierID, $result): bool
+    public function rejeteDossier($dossierID, $lastState): bool
     {
         /** @var SignatureConnecteur $signature */
         $signature = $this->getConnecteur('signature');
@@ -94,13 +111,15 @@ class FournisseurCommandeReceptionParapheur extends ActionExecutor
 
         $signature->effacerDossierRejete($dossierID);
 
-        $this->notify('rejet-iparapheur', $this->type, "Le document a été rejeté dans le parapheur : $result");
+        $message = 'Le document a été rejeté dans le parapheur : ' . $lastState;
         $this->getActionCreator()->addAction(
             $this->id_e,
             $this->id_u,
-            'rejet-iparapheur',
-            "Le document a été rejeté dans le parapheur : $result"
+            self::ACTION_NAME_REJECT,
+            $message,
         );
+        $this->notify(self::ACTION_NAME_REJECT, $this->type, $message);
+
         return true;
     }
 
@@ -114,7 +133,10 @@ class FournisseurCommandeReceptionParapheur extends ActionExecutor
         $signature = $this->getConnecteur('signature');
         $donneesFormulaire = $this->getDonneesFormulaire();
 
-        $info = $signature->getSignature($dossierID, false);
+        $document_element = 'commande';
+        $document_orignal_element = 'document_orignal';
+
+        $info = $signature->getSignature($dossierID);
         if (! $info) {
             $this->setLastMessage("La signature n'a pas pu être récupérée : " . $signature->getLastError());
             return false;
@@ -127,21 +149,23 @@ class FournisseurCommandeReceptionParapheur extends ActionExecutor
                 'signature.zip',
                 $signature->getDetachedSignature($info)
             );
-        }
+        } else {
+            $document_original_name = $donneesFormulaire->getFileName($document_element);
+            $document_original_data = $donneesFormulaire->getFileContent($document_element);
+            $filename = pathinfo($document_original_name, PATHINFO_FILENAME);
+            $extension = pathinfo($document_original_name, PATHINFO_EXTENSION);
 
+            if (!$donneesFormulaire->getFileName($document_orignal_element)) {
+                $donneesFormulaire->addFileFromData(
+                    $document_orignal_element,
+                    $document_original_name,
+                    $document_original_data
+                );
+            }
 
-        $originalDocumentName = $donneesFormulaire->getFileName('document_orignal');
-        if (!$originalDocumentName) {
-            $document_original_name = $donneesFormulaire->getFileName('commande');
-            $document_original_data = $donneesFormulaire->getFileContent('commande');
-            $donneesFormulaire->addFileFromData('document_orignal', $document_original_name, $document_original_data);
-        }
-        if ($info['document_signe']['document'] && !$originalDocumentName) {
-            $filename = substr($donneesFormulaire->getFileName('commande'), 0, -4);
-            $file_extension =  substr($donneesFormulaire->getFileName('commande'), -3);
-            $filename_signe = preg_replace("#[^a-zA-Z0-9_]#", "_", $filename) . "_signe." . $file_extension;
+            $filename_signe = preg_replace('#[^a-zA-Z0-9_]#', '_', $filename) . '_signe.' . $extension;
             $donneesFormulaire->addFileFromData(
-                'commande',
+                $document_element,
                 $filename_signe,
                 $signature->getSignedFile($info)
             );